+++ src/components-map/index.js
@@ -1,8 +1,0 @@
-import { buildForInline, build, buildForNamespaces, generateId } from './components-map';
-
-export {
-  buildForInline,
-  build,
-  buildForNamespaces,
-  generateId,
-};