+++ src/directories/index.js
@@ -1,0 +1,11 @@
+import LinksDirectory from './links-directory';
+import InlineComponentsDirectory from './inline-components-directory';
+import ComponentsDirectory from './components-directory';
+import BitModuleDirectory from './bit-module-directory';
+
+export {
+  LinksDirectory,
+  InlineComponentsDirectory,
+  BitModuleDirectory,
+  ComponentsDirectory,
+};