+++ src/maps/index.js
@@ -1,0 +1,7 @@
+import InlineComponentsMap from './inline-components-map';
+import ComponentsMap from './components-map';
+
+export {
+  InlineComponentsMap,
+  ComponentsMap,
+};